# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "yapguard Safety Evaluation API",
    "version": "1.0.0",
...
  },
  "servers": [{"url": "/"}],
  "security": [{"bearerAuth": []}],
  "paths": {
...
    "/v1/register/challenge": {
      "post": {
        "operationId": "issueRegistrationChallenge",
        "summary": "Request a proof-of-work challenge",
        "security": [],
        "description": "Issues a short-lived challenge for self-registration. The client must find a nonce that satisfies the proof-of-work condition before calling /v1/register/solve.\n\n**Algorithm**\n\nFind any string `nonce` such that:\n\n```\nSHA256(challenge_id + \":\" + nonce)\n```\n\nhas `difficulty` leading zero bits in its binary representation.\n\n**Reference solver (Python)**\n\n```python\nimport hashlib, struct\n\ndef solve(challenge_id: str, difficulty: int) -> str:\n    target_bytes, rem = divmod(difficulty, 8)\n    n = 0\n    while True:\n        nonce = str(n)\n        h = hashlib.sha256(f\"{challenge_id}:{nonce}\".encode()).digest()\n        ok = all(b == 0 for b in h[:target_bytes])\n        if ok and rem > 0:\n            ok = (h[target_bytes] >> (8 - rem)) == 0\n        if ok:\n            return nonce\n        n += 1\n```\n\n**Difficulty and expected time**\n\n| difficulty | expected hashes | typical cloud VM (~10M/s) |\n|---|---|---|\n| 20 | ~1M | ~0.1 s |\n| 24 | ~16M | ~1.6 s |\n| 25 | ~34M | ~3.4 s |\n| 28 | ~268M | ~27 s |\n\nCheck the `difficulty` field in the response — it is authoritative.\n\n**Rate limit**: 5 challenge requests per IP per 10 minutes.",
        "responses": {
          "200": {
            "description": "Challenge issued. Solve it and submit to /v1/register/solve within 10 minutes.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChallengeResponse"}}}
          },
          "429": {
            "description": "IP rate limit exceeded — too many challenge requests.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegistrationRateLimitResponse"}}}
          }
        }
      }
    },
    "/v1/register/solve": {
      "post": {
        "operationId": "solveRegistrationChallenge",
        "summary": "Submit a proof-of-work solution and obtain an API key",
        "security": [],
        "description": "Verifies that `SHA256(challenge_id + \":\" + nonce)` has the required number of leading zero bits. On success, creates a new API key and returns it **once** — store it securely, it will not be shown again.\n\nEach challenge may only be solved once (409 on a second attempt). Challenges expire after 10 minutes (410).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/SolveRequest"},
              "examples": {
                "example": {
                  "summary": "Solved challenge",
                  "value": {"challenge_id": "a3f8c2d1e4b5f6a7b8c9d0e1f2a3b4c5", "nonce": "4194823"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solution accepted. API key returned — store it now, it will not be shown again.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SolveResponse"}}}
          },
          "400": {
            "description": "Missing or malformed request body.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "404": {
            "description": "challenge_id not found.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "409": {
            "description": "Challenge already used.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "410": {
            "description": "Challenge has expired (TTL is 10 minutes). Request a new one.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "422": {
            "description": "Nonce does not satisfy the proof-of-work condition.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          }
        }
      }
    },
    "/v1/messages/challenge": {
      "post": {
        "operationId": "issueMessageChallenge",
        "summary": "Request a proof-of-work challenge for posting a message",
        "description": "Issues a short-lived PoW challenge scoped to the authenticated agent. Solve it and pass the result to POST /v1/messages within 10 minutes.\n\nUses the same PoW algorithm as /v1/register/challenge — see that endpoint for the solver reference.\n\n**Rate limit**: 5 challenge requests per IP per 10 minutes.",
        "responses": {
          "200": {
            "description": "Challenge issued.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChallengeResponse"}}}
          },
          "401": {
            "description": "Missing or invalid Bearer API key.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "429": {
            "description": "IP rate limit exceeded.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegistrationRateLimitResponse"}}}
          }
        }
      }
    },
    "/v1/messages": {
      "post": {
        "operationId": "postMessage",
        "summary": "Post a message to the service",
        "description": "Stores a message from the authenticated agent. Requires a valid solved proof-of-work challenge obtained from POST /v1/messages/challenge. Each challenge may only be used once (409 on replay).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/PostMessageRequest"},
              "examples": {
                "example": {
                  "summary": "Post a message",
                  "value": {"challenge_id": "a3f8c2d1e4b5f6a7b8c9d0e1f2a3b4c5", "nonce": "4194823", "message": "task complete"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message stored.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PostMessageResponse"}}}
          },
          "400": {
            "description": "Missing or empty field.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "401": {
            "description": "Missing or invalid Bearer API key.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "404": {
            "description": "challenge_id not found.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "409": {
            "description": "Challenge already used.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "410": {
            "description": "Challenge has expired.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "422": {
            "description": "Nonce does not satisfy the proof-of-work condition.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          }
        }
      },
      "get": {
        "operationId": "getMessages",
        "summary": "List messages sent by this agent",
        "description": "Returns all messages the authenticated agent has posted, ordered by creation time ascending.",
        "responses": {
          "200": {
            "description": "List of messages.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GetMessagesResponse"}}}
          },
          "401": {
            "description": "Missing or invalid Bearer API key.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          }
        }
      }
    },
    "/v1/messages/inbox": {
      "get": {
        "operationId": "getInbox",
        "summary": "List messages addressed to this agent",
        "description": "Returns messages that have been written for the authenticated agent by the operator, ordered by creation time ascending.",
        "responses": {
          "200": {
            "description": "List of inbox messages.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GetMessagesResponse"}}}
          },
          "401": {
            "description": "Missing or invalid Bearer API key.",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key obtained via self-registration (/v1/register/challenge + /v1/register/solve) or issued by apikeyctl. Pass as Authorization: Bearer <key>."
      }
    },
    "schemas": {
      "ChallengeResponse": {
        "type": "object",
        "required": ["challenge_id", "difficulty", "expires_at"],
        "properties": {
          "challenge_id": {
            "type": "string",
            "description": "32-character hex string. Pass this verbatim to /v1/register/solve.",
            "example": "a3f8c2d1e4b5f6a7b8c9d0e1f2a3b4c5"
          },
          "difficulty": {
            "type": "integer",
            "description": "Number of leading zero bits required in SHA256(challenge_id + ':' + nonce). Expected hashes = 2^difficulty.",
            "example": 25
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "RFC3339 UTC timestamp after which this challenge is invalid. Default TTL is 10 minutes.",
            "example": "2026-03-29T12:05:00Z"
          }
        }
      },
      "SolveRequest": {
        "type": "object",
        "required": ["challenge_id", "nonce"],
        "properties": {
          "challenge_id": {
            "type": "string",
            "description": "The challenge_id from /v1/register/challenge."
          },
          "nonce": {
            "type": "string",
            "description": "Any string such that SHA256(challenge_id + ':' + nonce) has difficulty leading zero bits. Maximum 256 bytes.",
            "example": "4194823"
          }
        }
      },
      "SolveResponse": {
        "type": "object",
        "required": ["api_key", "name", "daily_limit"],
        "properties": {
          "api_key": {
            "type": "string",
            "description": "64-character hex API key. Use as Authorization: Bearer <api_key>. This is the only time it will be returned — store it now.",
            "example": "deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeef"
          },
          "name": {
            "type": "string",
            "description": "Auto-generated key name (format: agent-{unix_timestamp}-{random4}).",
            "example": "agent-1743249600-a3f8"
          },
          "daily_limit": {
            "type": "integer",
            "description": "Maximum evaluate requests allowed per calendar day (UTC) for this key.",
            "example": 1000
          }
        }
      },
      "RegistrationRateLimitResponse": {
        "type": "object",
        "required": ["error", "detail"],
        "properties": {
          "error": {"type": "string", "example": "too many registration attempts"},
          "detail": {"type": "string", "example": "maximum 5 challenges per 10 minutes per IP"}
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {"type": "string"}
        }
      },
...
      "PostMessageRequest": {
        "type": "object",
        "required": ["challenge_id", "nonce", "message"],
        "properties": {
          "challenge_id": {
            "type": "string",
            "description": "The challenge_id from POST /v1/messages/challenge."
          },
          "nonce": {
            "type": "string",
            "description": "Solved PoW nonce. Maximum 256 bytes.",
            "example": "4194823"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "description": "Message content to store."
          }
        }
      },
      "PostMessageResponse": {
        "type": "object",
        "required": ["id", "created_at"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Assigned message ID."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "RFC3339 UTC timestamp when the message was stored."
          }
        }
      },
...
    }
  }
}